TFM-Code/requirements.txt
pandas==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
plotly==5.24.1
streamlit==1.38.0
typing-extensions==4.5.0

prophet==1.1.5
cmdstanpy==1.2.4
statsmodels==0.14.2

tensorflow-cpu==2.13.1
python-dateutil==2.9.0.post0
# pydantic  <-- ELIMINADO